pyproject.toml
[tool.poetry]
name = "zaim-io-tool"
version = "0.1.0"
description = ""
authors = ["showchan33 <showchan3333@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
playwright = "^1.49.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
